refactor(models): split TorchGeo wrappers by family
